Split validation tests into 2 files
 - 1 for the basic schema adherence
 - 1 for the extra validation bits
