+++ src/test/java/org/neo4j/importer/v1/ImportSpecificationDeserializerTest.java
@@ -588,642 +588,4 @@ class ImportSpecificationDeserializerTest {
                 .hasMessageContainingAll(
                         "1 error(s)", "0 warning(s)", "$.actions[0].name: does not match the regex pattern \\S+");
     }
-
-    @Test
-    void fails_if_source_name_is_duplicated_with_node_target() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "duplicate",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "nodes": [{
-                            "name": "duplicate",
-                            "source": "duplicate",
-                            "labels": ["Label1", "Label2"],
-                            "write_mode": "create",
-                            "properties": [
-                                {"source_field": "field_1", "target_property": "property1"},
-                                {"source_field": "field_2", "target_property": "property2"}
-                            ]
-                        }]
-                    }
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "Name \"duplicate\" is duplicated across the following paths: $.sources[0].name, $.targets.nodes[0].name");
-    }
-
-    @Test
-    void fails_if_source_name_is_duplicated_with_rel_target() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "duplicate",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "relationships": [{
-                            "name": "duplicate",
-                            "source": "duplicate",
-                            "type": "TYPE",
-                            "start_node": {
-                                "label": "Label1",
-                                "key_properties": [
-                                    {"source_field": "field_1", "target_property": "property1"}
-                                ]
-                            },
-                            "end_node": {
-                                "label": "Label2",
-                                "key_properties": [
-                                    {"source_field": "field_2", "target_property": "property2"}
-                                ]
-                            }
-                        }]
-                    }
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "Name \"duplicate\" is duplicated across the following paths: $.sources[0].name, $.targets.relationships[0].name");
-    }
-
-    @Test
-    void fails_if_source_name_is_duplicated_with_custom_query_target() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "duplicate",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "queries": [{
-                            "name": "duplicate",
-                            "source": "duplicate",
-                            "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                        }]
-                    }
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "Name \"duplicate\" is duplicated across the following paths: $.sources[0].name, $.targets.queries[0].name");
-    }
-
-    @Test
-    void fails_if_source_name_is_duplicated_with_action() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "duplicate",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "queries": [{
-                            "name": "my-target",
-                            "source": "duplicate",
-                            "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                        }]
-                    },
-                    "actions": [{
-                        "name": "duplicate",
-                        "type": "http",
-                        "method": "get",
-                        "url": "https://example.com"
-                    }]
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "Name \"duplicate\" is duplicated across the following paths: $.sources[0].name, $.actions[0].name");
-    }
-
-    @Test
-    void fails_if_node_target_does_not_refer_to_existing_source() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "nodes": [{
-                    "name": "a-target",
-                    "source": "incorrect-source-name",
-                    "labels": ["Label1", "Label2"],
-                    "write_mode": "create",
-                    "properties": [
-                        {"source_field": "field_1", "target_property": "property1"},
-                        {"source_field": "field_2", "target_property": "property2"}
-                    ]
-                }]
-            }
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.nodes[0] refers to the non-existing source \"incorrect-source-name\". "
-                                + "Possible names are: \"a-source\"");
-    }
-
-    @Test
-    void fails_if_relationship_target_does_not_refer_to_existing_source() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "relationships": [{
-                    "name": "a-target",
-                    "source": "incorrect-source-name",
-                    "type": "TYPE",
-                    "start_node": {
-                        "label": "Label1",
-                        "key_properties": [
-                            {"source_field": "field_1", "target_property": "property1"}
-                        ]
-                    },
-                    "end_node": {
-                        "label": "Label2",
-                        "key_properties": [
-                            {"source_field": "field_2", "target_property": "property2"}
-                        ]
-                    }
-                }]
-            }
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.relationships[0] refers to the non-existing source \"incorrect-source-name\". "
-                                + "Possible names are: \"a-source\"");
-    }
-
-    @Test
-    void fails_if_custom_query_target_does_not_refer_to_existing_source() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "queries": [{
-                    "name": "a-target",
-                    "source": "incorrect-source-name",
-                    "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                }]
-            }
-        }
-"""
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.queries[0] refers to the non-existing source \"incorrect-source-name\". "
-                                + "Possible names are: \"a-source\"");
-    }
-
-    @Test
-    void fails_if_node_target_depends_on_non_existing_target_or_action() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "nodes": [{
-                    "name": "a-target",
-                    "source": "a-source",
-                    "depends_on": "incorrect-dependent",
-                    "labels": ["Label1", "Label2"],
-                    "write_mode": "create",
-                    "properties": [
-                        {"source_field": "field_1", "target_property": "property1"},
-                        {"source_field": "field_2", "target_property": "property2"}
-                    ]
-                }]
-            }
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.nodes[0] depends on a non-existing action or target \"incorrect-dependent\"");
-    }
-
-    @Test
-    void fails_if_relationship_target_depends_on_non_existing_target_or_action() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "relationships": [{
-                    "name": "a-target",
-                    "source": "a-source",
-                    "depends_on": "incorrect-dependent",
-                    "type": "TYPE",
-                    "start_node": {
-                        "label": "Label1",
-                        "key_properties": [
-                            {"source_field": "field_1", "target_property": "property1"}
-                        ]
-                    },
-                    "end_node": {
-                        "label": "Label2",
-                        "key_properties": [
-                            {"source_field": "field_2", "target_property": "property2"}
-                        ]
-                    }
-                }]
-            }
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.relationships[0] depends on a non-existing action or target \"incorrect-dependent\"");
-    }
-
-    @Test
-    void fails_if_custom_query_target_depends_on_non_existing_target_or_action() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "queries": [{
-                    "name": "a-target",
-                    "source": "a-source",
-                    "depends_on": "incorrect-dependent",
-                    "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                }]
-            }
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.queries[0] depends on a non-existing action or target \"incorrect-dependent\"");
-    }
-
-    @Test
-    void fails_if_action_depends_on_non_existing_target_or_action() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "queries": [{
-                    "name": "a-target",
-                    "source": "a-source",
-                    "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                }]
-            },
-            "actions": [{
-                "name": "an-action",
-                "depends_on": "incorrect-dependent",
-                "type": "http",
-                "method": "get",
-                "url": "https://example.com"
-            }]
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.actions[0] depends on a non-existing action or target \"incorrect-dependent\".");
-    }
-
-    @Test
-    void fails_if_relationship_refers_to_a_non_existing_node_target_for_start() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-        {
-            "sources": [{
-                "type": "bigquery",
-                "name": "a-source",
-                "query": "SELECT id, name FROM my.table"
-            }],
-            "targets": {
-                "relationships": [{
-                    "name": "a-target",
-                    "source": "a-source",
-                    "type": "TYPE",
-                    "start_node_reference": "incorrect-reference",
-                    "end_node": {
-                        "label": "Label2",
-                        "key_properties": [
-                            {"source_field": "field_2", "target_property": "property2"}
-                        ]
-                    }
-                }]
-            }
-        }
-        """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.relationships[0].start_node_reference refers to a non-existing node target \"incorrect-reference\".");
-    }
-
-    @Test
-    void fails_if_relationship_refers_to_a_non_existing_node_target_for_end() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "a-source",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "relationships": [{
-                            "name": "a-target",
-                            "source": "a-source",
-                            "type": "TYPE",
-                            "start_node": {
-                                "label": "Label1",
-                                "key_properties": [
-                                    {"source_field": "field_1", "target_property": "property1"}
-                                ]
-                            },
-                            "end_node_reference": "incorrect-reference"
-                        }]
-                    }
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.relationships[0].end_node_reference refers to a non-existing node target \"incorrect-reference\".");
-    }
-
-    @Test
-    void fails_if_direct_dependency_cycle_is_detected() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "a-source",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "relationships": [{
-                            "name": "a-target",
-                            "depends_on": "a-target",
-                            "source": "a-source",
-                            "type": "TYPE",
-                            "start_node": {
-                                "label": "Label1",
-                                "key_properties": [
-                                    {"source_field": "field_1", "target_property": "property1"}
-                                ]
-                            },
-                            "end_node": {
-                                "label": "Label2",
-                                "key_properties": [
-                                    {"source_field": "field_2", "target_property": "property2"}
-                                ]
-                            }
-                        }]
-                    }
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "A dependency cycle has been detected",
-                        "\"a-target\" ($.targets.relationships[0]) depends on \"a-target\" ($.targets.relationships[0])");
-    }
-
-    @Test
-    void fails_if_longer_dependency_cycle_is_detected() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "a-source",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "relationships": [{
-                            "name": "a-relationship-target",
-                            "depends_on": "an-action",
-                            "source": "a-source",
-                            "type": "TYPE",
-                            "start_node": {
-                                "label": "Label1",
-                                "key_properties": [
-                                    {"source_field": "field_1", "target_property": "property1"}
-                                ]
-                            },
-                            "end_node": {
-                                "label": "Label2",
-                                "key_properties": [
-                                    {"source_field": "field_2", "target_property": "property2"}
-                                ]
-                            }
-                        }],
-                        "queries": [{
-                            "name": "a-query-target",
-                            "source": "a-source",
-                            "depends_on": "a-relationship-target",
-                            "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                        }]
-                    },
-                    "actions": [{
-                        "name": "an-action",
-                        "depends_on": "a-query-target",
-                        "type": "http",
-                        "method": "get",
-                        "url": "https://example.com"
-                    }]
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "A dependency cycle has been detected",
-                        "\"a-relationship-target\" ($.targets.relationships[0]) depends on \"an-action\" ($.actions[0])",
-                        "\"an-action\" ($.actions[0]) depends on \"a-query-target\" ($.targets.queries[0])",
-                        "\"a-query-target\" ($.targets.queries[0]) depends on \"a-relationship-target\" ($.targets.relationships[0])");
-    }
-
-    @Test
-    void does_not_report_cycles_if_names_are_duplicated() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "a-source",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "relationships": [{
-                            "name": "a-target",
-                            "depends_on": "an-action",
-                            "source": "a-source",
-                            "type": "TYPE",
-                            "start_node": {
-                                "label": "Label1",
-                                "key_properties": [
-                                    {"source_field": "field_1", "target_property": "property1"}
-                                ]
-                            },
-                            "end_node": {
-                                "label": "Label2",
-                                "key_properties": [
-                                    {"source_field": "field_2", "target_property": "property2"}
-                                ]
-                            }
-                        }],
-                        "queries": [{
-                            "name": "a-target",
-                            "source": "a-source",
-                            "depends_on": "an-action",
-                            "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                        }]
-                    },
-                    "actions": [{
-                        "name": "an-action",
-                        "depends_on": "a-target",
-                        "type": "http",
-                        "method": "get",
-                        "url": "https://example.com"
-                    }]
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "Name \"a-target\" is duplicated across the following paths: $.targets.relationships[0].name, $.targets.queries[0].name");
-    }
-
-    @Test
-    void does_not_report_cycles_if_depends_on_are_dangling() {
-        assertThatThrownBy(() -> deserialize(new StringReader(
-                        """
-                {
-                    "sources": [{
-                        "type": "bigquery",
-                        "name": "a-source",
-                        "query": "SELECT id, name FROM my.table"
-                    }],
-                    "targets": {
-                        "relationships": [{
-                            "name": "a-target",
-                            "depends_on": "an-action",
-                            "source": "a-source",
-                            "type": "TYPE",
-                            "start_node": {
-                                "label": "Label1",
-                                "key_properties": [
-                                    {"source_field": "field_1", "target_property": "property1"}
-                                ]
-                            },
-                            "end_node": {
-                                "label": "Label2",
-                                "key_properties": [
-                                    {"source_field": "field_2", "target_property": "property2"}
-                                ]
-                            }
-                        }],
-                        "queries": [{
-                            "name": "a-query-target",
-                            "source": "a-source",
-                            "depends_on": "invalid-depends-on",
-                            "query": "UNWIND $rows AS row CREATE (n:ANode) SET n = row"
-                        }]
-                    },
-                    "actions": [{
-                        "name": "an-action",
-                        "depends_on": "a-target",
-                        "type": "http",
-                        "method": "get",
-                        "url": "https://example.com"
-                    }]
-                }
-                """
-                                .stripIndent())))
-                .isInstanceOf(InvalidSpecificationException.class)
-                .hasMessageContainingAll(
-                        "1 error(s)",
-                        "0 warning(s)",
-                        "$.targets.queries[0] depends on a non-existing action or target \"invalid-depends-on\"");
-    }
 }
